Updated the technique for resolving the physical filename.
filename() doesn't give you the path to the file as requested...it gives
you how much of the path to the requested file actually exists.  This
leads to a false negative on the 404 test if there is a file which is a
substring of the requested one. (uri: /foo/bar/baz file: /foo/bar)